requirements.txt
# 微信智能浇水上报系统 - 依赖包
# WeChat Smart Watering Reporting System - Dependencies

# =========================
# Web框架
# =========================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# =========================
# 数据库
# =========================
sqlalchemy==2.0.25; python_version < "3.13"
sqlalchemy>=2.0.37,<3; python_version >= "3.13"
pymysql==1.1.0
alembic==1.13.1

# =========================
# 微信SDK
# =========================
wechatpy==1.8.18

# =========================
# 大模型
# =========================
openai==1.10.0
zhipuai==2.1.5.20250625

# =========================
# 配置管理
# =========================
pyyaml>=6.0.2,<7
python-dotenv==1.0.0
pydantic==2.5.3; python_version < "3.13"
pydantic>=2.10.0,<3; python_version >= "3.13"
pydantic-settings==2.1.0; python_version < "3.13"
pydantic-settings>=2.6.1,<3; python_version >= "3.13"

# =========================
# 工具类
# =========================
httpx==0.26.0
python-dateutil==2.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# =========================
# 日志
# =========================
loguru==0.7.2

# =========================
# 测试
# =========================
pytest==7.4.4
pytest-asyncio==0.23.3
